requirements.txt
google-api-python-client
gspread